+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-11 19:49:35
-// Commit: feat(network): integrate custom telemetry logger in network
+// Updated at: 2026-08-11 20:49:35
+// Commit: test(network): implement widget integration tests for network
 
 class networkComponentState {
-  final String commitRef = '3545';
-  final String lastMessage = 'feat(network): integrate custom telemetry logger in network';
-  final DateTime timestamp = DateTime.parse('2026-08-11T19:49:35+05:00');
+  final String commitRef = '3547';
+  final String lastMessage = 'test(network): implement widget integration tests for network';
+  final DateTime timestamp = DateTime.parse('2026-08-11T20:49:35+05:00');
 }